Update toolchains to 9.2-2019.12 and Linaro components to 19.06
Updates toolchains (TC) and Linaro components (LC). For the case of
TC, besides updating from 8.3-2019.03 to 9.2-2019.12, the new TC
are 'baremetal' (aarch64-none-elf and arm-none-eabi). For the case of
LC, the upgrade is from 18.04 to 19.06. Besides the upgrades, some old
comments were removed, turning the script a bit more readable.

Change-Id: If45555b2d1dbcaae51a37ed9abb1b441a2000110
Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>