install linaro toolchain aarch64-linux-gnu version 6.2-2016-11
Besides installing the Linaro toolchain, it does some refactor of
toolchain installation, turning into two functions: one install Linaro
toolchains and a second the ARM ones. For the moment, the only Linaro
Toolchain considered is aarch64-linux-gnu, but code allows to easily
include another one in the future. The new toolchain is included into
PATH, so no need to provide absolute path but if it is strictly
needed, one can use it direclty:

	<WORKSPACE>/gcc-linaro-6.2.1-2016.11-x86_64_aarch64-linux-gnu/bin/aarch64-linux-gnu-gcc

or through the common environment variable CROSS_COMPILE

   export CROSS_COMPILE=<WORKSPACE>/gcc-linaro-6.2.1-2016.11-x86_64_aarch64-linux-gnu/bin/aarch64-linux-gnu-

Change-Id: Ibc109aff4e3d9c553ad107595511573bb318a63b
Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>