Migrate Arm CI TF L1 (tf-gerrit-tforg-l1) level job
This is the start of a series of patches with the aim to converge CI
approaches between TF-A Arm [1] and Open [2] CIs. Taking the best of
both worlds, this patch migrates from Arm CI the way it triggers the
rest of the jobs and define test groups, all done in
'tf-gerrit-tforg-l1' job. These test groups are further
splitted by 'tf-ci-gateway' job, and finally sending
a single test configuration to 'tf-a-builder' job, the
package builder, where user changes are taken into account and
compiled. The end of the CI flow is 'post-build-lava' job, in charge
of posting to LAVA [3].

As described above, this is a not full Arm CI migration, but a
converge of both CIs. The decision to take the existing Open CI
'tf-a-builder' and 'post-build-lava' jobs as package builder and LAVA
submitter instead of Arm CI jobs 'tf-worker' and 'tf-build-for-lava'
is mainly for *scalability* and *maintainability* reasons. In this
scenario, both reasons nails down to the main job of the CI: build and
test, where the former is done at *isolation* under a containerized
environment, without being tied to master node HW capacity. In other
words, while Arm CI uses master node for all CI computation, from
triggering up to building, the new proposal only uses master node
for triggering & splitting the work (the test groups), then the
building piece is offloaded into a container
node (docker-amd64-tf-a-bionic) [4].

[1] https://jenkins.oss.arm.com/ (this is an internal ci, cannot be
acccess from outside arm)
[2] https://ci.trustedfirmware.org/
[3] https://tf.validation.linaro.org/scheduler/
[4] https://git.trustedfirmware.org/ci/dockerfiles.git/tree/bionic-amd64-tf-a-build/Dockerfile

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I764cc2a4ba381bb83532033e5a950a1dddf6a2f6