tf-static-checks: introduce job
The proposed job (tf-static-checks) effectively superseed
'tf-a-static-checks.yaml', where besides the different names, the new
jobs is formatted similar as the recent migrated jobs and also
introduce a summary page as seen in [1].

This patch also fixes tf-main's static check job call, which now calls
the tf-static-checks job. Also, the static check code at
tf-a-builder is removed, leaving the latter job with the only purpose
of building (and nothing more).

[1] https://ci.staging.trustedfirmware.org/job/lsandov1-tf-static-checks/19/

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: Ic605ec8fe84445b080aa0dd38d4f85d289411656