tf-a-builder: turn build as failure if LAVA job fails
Currently, the result (success or failure) of the tf-a-builder job
depends only on the execution of the build phase (builders.sh) without
taking into account the LAVA job result. This commit, turns
tf-a-builder's status into 'failure' if builds passes and LAVA job
fails. There are now two tf-a-builder's status representing:

    1. Success:  Build job succeeded and LAVA job succeeded
    3. Failure:  Build job failed or LAVA job failed

A LAVA job is considered as failure if ANY of the LAVA jobs suite
fails. For example, this is considered a failure because there is at
least one failure (fail)

    * lava.job [fail]
    * lava.boot-fvp [fail]
    * lava.boot-fvp-main [fail]
    * lava.run-fvp [fail]
    * lava.fvp-version [pass]
    * lava.http-download [pass]
    * lava.http-download [pass]
    * lava.http-download [pass]
    * lava.http-download [pass]
    * lava.http-download [pass]
    * lava.validate [pass]

Tested at staging [1]

[1] https://ci.staging.trustedfirmware.org/job/lsandov1-tf-main/34/

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I2c39a8cd59f5b5a26b8195ac45c4b5949ddf170c