build discarder: adjust days-to-keep and remove unnecessary num-to-keep
Besides aligning to build-discarder ARM job values, it keeps simpler
the logic for TF-A CI build-discarder: for most jobs, expect tf-main,
the only parameter to consider is days-to-keep and it is adjusted for
realistic values (14 instead of 180), reducing job-storage footprint
considerably.

This patch also fixes the problem observed in current tf-main jobs:
last-week tf-main builds pointed to non-existent tf-ci-gateway jobs
because the num-to-keep limit reached first than days-to-keep, so
these jobs were deleted which is an unwanted behavior.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I65a404d6994b498abafcc1ae5896ce96df49736c