builders.sh: include tf-l3-code-coverage and one test desc in the blocklist
One test group, 'tf-l3-code-coverage', and one particular test
description, see below, are now part of the blocklist so these will
not be considered for the full CI process.

The reason for the former is that Arm needs to shared a library
introduced by [1], and for the latter, Open CI needs to support the
same logic as 'expect' to monitor and interact UART logs.

Test description:
fvp-tbb-mbedtls-upcounter,fvp-fwu:fvp-tftf.fwu-aemv8a.invalid_nvcounter-debug

[1] https://git.trustedfirmware.org/ci/tf-a-ci-scripts.git/commit/?id=024efd51f025b8446e87c9adf99bd770dc769d99

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I9a9c81228dd6ad7b1fbcc2cee2a3153c53a1ed06