squad.sh: timeout lavacli-job-wait command
There are cases where the LAVA job (callee) hangs indefinitely,
causing the CI job tf-a-builder (caller) to hang, affecting current
and future jobs in the Open CI [1].

The proposed approach runs the lavacli-job-wait command through the
timeout command with a max duration of 1 hour (3600); if LAVA job runs
longer that 1 hour, command timeouts and exits. An artificial lava.log
is created, with the following  content

    Stopped monitoring LAVA JOB XXX after 3600 seconds, likely stuck or timeout too short?

where XXX is of course correctly populated and tf-a-builder job is
failed.

This can be considered a workaround and should be fixed either
by fixing current LAVA jobs timeouts, CI jobs timeouts and/or LAVA job
monitors, but for the time being, this is an effective approach which
avoids undesired effects as in [1].

[1] https://projects.linaro.org/browse/TFC-43

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I94895f58d90fcee8f67c288027549b159d1cdb42