squad.sh: use lavacli-jobs-show output as a LAVA job pass/fail criteria
The 'lavacli results' commands indicates cases/suites results but what
we need to know is the health & state of the job, these two provided
by the 'lavacli jobs show' command, so use this command instead of the
current one.

A successful job is considered if health value is 'Complete' AND state
is 'Finished'. Any other combination is considered as a failure.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I9538dc599d56d67d3bf35dd2c7e3195cc52d6031