tf-a-builder: remove num-to-keep
The testing volume taking place at Open CI leads to thousands of
tf-a-build jobs being launched in a short period of time, leading to
useless tf-ci-gateway tables as in [1], table created a approx. 5 days
ago but listed tf-a-build jobs are not found anymore (404). This
commit remove the num-to-keep restriction, effectively leaving the
jobs for two weeks, as indicated in the days-to-keep parameter.

[1] https://ci.trustedfirmware.org/job/tf-ci-gateway/3793/

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I6613c096f703cb5f7884456f0b82ffe38f3d9725