tf-static-checks: include REPO_UNDER_TEST choice parameter
Currently the tf-static-checks job can only (static)check the core project
trusted-firmare-a, but job must be to check at least the tf-a-tests
repository. This commit enables a new choice parameter, REPO_UNDER_TEST,
defaulting to trusted-firmware-a but including the tf-a-tests
repository in the choice list.

Enabling another choice project, implies repository cloning and
including gerrit related parameters, TFTF_GERRIT_*, so this changes
are part of the commit.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I468124a07542f72e2c8a16191b4ee701f8d98827