builders.sh: block two L1 test descriptions
Currently, all L1 (and L2, because L1 is triggered by L2 also) jobs
are failing due to two LAVA jobs failing, the latter triggered by two
tf-a-builder jobs which are part of 'Run boot tests' L1 multijob.

This is a workaround to avoid production failures while the root issue
is found and must be reverted once correct fix is found/merged.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: Iba119d2716c0169ee463d6c86c518aa658d54532