tf-a-builder.yaml: resubmit a juno LAVA job in case of failure
A Juno LAVA job may fail due to possible infrastructure errors* [1], not
related to user changes. These failures are most of the time
transient, related to a particular board itself [2] and not the LAVA
framework. Failures of these type should not impact gerrit impact
gerrit votes negatively.

Although LAVA resubmits automatically the job in case of
infrastructure problem, the latter job is injected with binaries
expected to work (without user changes), making it useless for jenkins
to reuse it, so there is a need to resubmit it.

* In fact, any other HW device can fail with the same reason but TF-A
only test Juno devices.

[1] https://projects.linaro.org/browse/TFC-76
[2] https://projects.linaro.org/browse/LSS-2128

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I2ba8bcab2fcf719ec7373cc4943acce7f53f808d