tf-gerrit-tforg-l1: use correct refname for commitlint
`GERRIT_BRANCH` gives the branch name that the patch is based on, as
opposed to the refname of the commit under test. This change replaces
the refname with the actual commit identifier, and instead uses the
Gerrit branch to detect the merge base.

The current setup is causing false positives, as the commit under test
is instead the tip of the branch.

Change-Id: I16e2f34a7e9e605c8ef228779414df8c546994fa
Signed-off-by: Chris Kay <chris.kay@arm.com>