squad.sh: increase wait-time when querying LAVA job ID
Once SQUAD submits a LAVA job, it does not immediately submit it to LAVA,
instead, it goes into a internal queue and can take several
minutes before getting submitted. To take into account the latter
latency, the wait-time is increased to 30 seconds, so the wait-for-ID
loop can take at most 10x30 seconds (5 minutes), enough time for the
job to be submitted.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I6ff294dde43be594cac93ca18bb4c4cf81ebb5ef