tf-a-builder.yaml: include QA_SERVER_VERSION parameter
Commonly, this parameter is populated/passed from the trigger
job (L1/L2) to tf-a-builder, but in case of a parametrized build from
tf-ci-gateway [1], this is not the case, so include into tf-a-builder
job. Resulting LAVA jobs from parameterized builds will be tracked by
SQUAD at [2] with the ID corresponding to the Jenkins Build ID, which
i the ID of the particular tf-a-build ID, i.e. [3]. This change would
allow LAVA jobs to be submitted (via SQUAD) from [1].

[1] https://ci.trustedfirmware.org/view/TF-A/job/tf-ci-gateway/build?delay=0sec
[2] https://qa-reports.linaro.org/tf/tf-a/
[3] https://qa-reports.linaro.org/tf/tf-a/build/48232/

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I51275083369bbbc2de70642d895073f6368a5860