tf-a-builder.yaml: generate individual code coverage reports
Parent commit introduces the script to split the coverage traces from
the LAVA log, and this commit calls it and use the reporting tool [1]
to create code coverage individual reports.

Patch has been tested in staging [2] where most of the reports could
not be created due [3], however there are two successful ones.

This work only covers reporting per build, but the merge of all
individual reports to create a single coverate report is still pending
because scripts are not yet public [4].

[1] https://gitlab.arm.com/tooling/qa-tools/-/tree/master/coverage-tool/coverage-reporting
[2] https://ci.staging.trustedfirmware.org/view/lsandov1-TF-A/job/lsandov1-tf-ci-gateway/2600/
[3] https://gitlab.arm.com/tooling/qa-tools/-/issues/1
[4] https://git.trustedfirmware.org/ci/tf-a-ci-scripts.git/tree/script/test_definitions.sh#n8

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I64d8d8870eb145264ddb0104c3847e4f002667b7