tf-a-builder.yaml: smart clone qa-tools repository
While reverting [1] would be the fastest way to re-enable the qa-tools
repository [2], this is too expensive in terms of CI latency: repository
is cloned per tf-a-builder job but ultimately it is only needed when
running code coverage test descriptions, i.e. tf-l3-code-coverage/*.

This commit clones the repository only when code coverage reports are
required.

[1]
https://git.trustedfirmware.org/ci/tf-a-job-configs.git/commit/?id=2c7f7aedc03ebb24ed072685e371a027de22be53
[2] https://git.gitlab.arm.com/tooling/qa-tools.git

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I9687253762f5b7765538712c2ea629c442ff4423