Introduce a global share repository mechanism
This is a full refactor of the way jobs fetch git data: with this
approach, only the top level job, the upstream job, would set a share
folder populated with all required repositories, including
trusted-firmware-a, tf-a-tests, tf-a-job-configs and tf-a-ci-scripts,
ultimately consumed by the builder job, tf-a-builder. This would
reduce considerably the CI build times, avoiding multiple clones per
job thus overloading the TF git servers as reported in [1].

The way it works is simple: the top level job sets the SHARE_FOLDER
parameter based on its name and number on top of the share
volume (/srv/shared/<job name>/<job number>) then it calls the clone
script (clone.sh), which in turn it fetches the repositories mentioned
above. Jobs triggered on behalf of the latter, share the same
SHARE_FOLDER value, and these in turn also call the clone script, but
in this case, the script detects that the folder is already populated so
its role is to simply copy the repositories into the job's
workspace. As seen, all jobs work with repositories on their own
workspace, which are just copies of the share folder, so there is no
change of a race condition, i.e every job works with its own copy. The
worst case scenario is where the down-level job, tf-a-builder, uses its
default SHARE_FOLDER value, in this case, it would simply clone its
own repositories without reusing any file however the current approach
prevents the latter unless the job is triggered manually from the
builder job itself.

Besides the git-cloning strategy presented, the change implied moving
all of jobs to 'docker-amd64-tf-a-bionic' container node, instead of
'master', the former having access to the docker share volume.

[1] https://linaro.atlassian.net/browse/TFC-20

Signed-off-by: Leonardo Sandoval <leonardo.sandoval@linaro.org>
Change-Id: I001bc1148cde4b1952161d84aa3d2e089a4281a2