visualizations: Add visualization jobs
This change adds a Source Lines of Code (SLoC) visualization job and a
test category visualization job.

The SLoC visualization job generates a stacked histogram representing
the lines of code in each programming language for a given list of
directories. This job takes an input parameter `DIRECTORIES` which
should consist of a list of space-separated directories to be included
in the visualization, where statistics are generated for each top-level
subdirectory. An example output from this job is shown [here][1].

The test category visualization job generates a stacked histogram
representing the number of test configurations enabled at each CI level.
This job takes an input parameter `FILTER` which represents a string
that must be included in the job or group name for the job to be
included in the visualization. An example output from this job is shown
[here][2].

[1]: https://ci-builds.staging.trustedfirmware.org/static-files/58yOtzp9JwWBVtvHrSGl_yEmefYiT6B3eRN-uO_56mExNjYxNDM1OTMxOTU3OjU6Q0pLYXk6am9iL2Nqa2F5LXRmLWEtc2xvYy12aXN1YWxpemF0aW9uLzIzL2FydGlmYWN0/output.png
[2]: https://ci-builds.staging.trustedfirmware.org/static-files/egmlKGYCpXtG-7Q-dXFrviEt_xqwMx83bM679VJURIIxNjYxNDM1OTYxMTc4OjU6Q0pLYXk6am9iL2Nqa2F5LXRmLWEtdGVzdC1jYXRlZ29yeS12aXN1YWxpemF0aW9uL2xhc3RTdWNjZXNzZnVsQnVpbGQvYXJ0aWZhY3Q=/output.png

Change-Id: I81849b77e18bb42109ac3bca2fc60f4482fb2fea
Signed-off-by: Chris Kay <chris.kay@arm.com>
Co-authored-by: Weronika Wiesiolek <weronika.wiesiolek@arm.com>