tf-a-builder: update mbedtls to 2.28
Upgrade to the latest and greatest 2.x release of Mbed TLS library
(i.e. v2.28.0) to take advantage of their bug fixes.

Note that the Mbed TLS project published version 3.x some time
ago. However, as this is a major release with API breakages, upgrading
to 3.x might require some more involved changes in TF-A, which we are
not ready to do. We shall upgrade to mbed TLS 3.x after the v2.7
release of TF-A.

See the companion TF-A patch here:
https://git.trustedfirmware.org/TF-A/trusted-firmware-a.git/commit/?h=integration&id=a93084be95634b66b917f1c8baf403067dc75c5d

Change-Id: I5b6f2db76aaf739060df8fcf52d87e4651a1267e
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>