tf-ci-gateway, tf-a-builder: Fix false positive Jenkins test result
There is a situation where a TF-A Jenkins job is reporting success when
an underlying test is failing and reported as failing in LAVA. [1]
After confirming with LAVA engineer, the job completed doesn't mean test
pass. It only means that the job ran properly without any infrastructure
issues (connection dropped, serial broken during execution etc).
We need to check the test results.

[1]: https://ci.trustedfirmware.org/view/TF-A/job/tf-a-ci-gateway/11258/

Signed-off-by: Arthur She <arthur.she@linaro.org>
Change-Id: I7522077ac871314e42e523c997b6d81d9001b775