squad.sh: Run curl with --fail option to properly detect REST errors
Currently, we're seeing situation when e.g. REST API requests leads to 404
error, it's not detected as an error on the script level, instead HTML
content of the error page is used as an REST API response, leading to
more request errors down the road (which are still not detected).
The --fail option causes curl to exit with non-0 exit code in this case,
to communicate the error to calling script (which runs with set -e).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I068ab70f02b218addbeae3140e32b6cd48d5a9c7