tf-a-builder.yaml: Update to mbedtls 2.28.1
In anticipation of the next Trusted Firmware release update to the
newest Mbed TLS library.

Note that the Mbed TLS project publised version 3.x some time ago.
However, as this is a major release with API breakages, upgrading to
this one might require some more involved changes in TF-A, which we
are not ready to do. We shall upgrade to Mbed TLS 3.x after the v2.8
release of TF-A.

Also update the URL to mbedtls repository on Github -
https://github.com/ARMmbed/mbedtls redirects to
https://github.com/Mbed-TLS/mbedtls, so we might as well directly
use the latter.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: I911531e9e3157c70406c07a9b3cb85e723d95a3a