tf-a-eclair-delta: Job to perform patch analysis using ECLAIR
Perform MISRA analysis using ECLAIR tool, for a patch (git commit).
Patch analysis is performed by running full analysis for base commit,
the for commit, then by using ECLAIR's "diff" capabilities. A single
TF-A configuration (fvp-default) is used so far, both out of concern
of system load and interpretation of the results (in the common case
of changes to generic code, the same issue would be shown again and
again, leading to unwieldy reports).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: If1a1d256a1a0f7f5774bf54f95bd452d82a29161