tf-a-eclair-delta: Throttle using "eclair" category
The "eclair" category set at 5 (full number of available ECLAIR licenses
as of now). Note that this category is disjoint with "eclair-daily" as
used by tf-a-eclair-daily (set at 3). The idea is to prevent take up
of all the available license by the long-running daily job, keeping
a couple available to the delta job, if a random patch or a couple
comes by. But note that if a full series comes at the same time,
this disjointness may lead to a real job error due to unavailable
licenses. We have a second line of handling this via retries in the
detachLicense script, but up to a limit. Overall, if the errors like
above will actually happen, we may drop wishful "optimizations" like
above and use the single category.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I34b2475bf37ed0dfdf6adb83cdbed292937667fb