log-splitter: update multiplexing logic of lava.log
It is assumed that the model supports uart0 to uart3 terminals,
with uart0 always being the primary uart.

However, in some CSS models, the primary uart may not always be
uart0, it can be any of the terminals that the model supports.
Also, lava-uart0 to lava-uart3 are very conventional names used
by the script, so they were renamed according to the terminal
namespace.

Also, updated builder job to achive necessary lava files.

Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>
Change-Id: I03c52e04fd6bce51ebd2df6123e7a61b0c4eae91