feat: add TFTF main job
This change adds the `tf-a-tftf-main` OpenCI job, which runs an
extensive suite of tests to validate the sanity of the Trusted
Firmware-A Test Framework project.

Contrary to the TF-A main job, this job runs on the TF-A `master` branch
and includes additional tests relevant only to TFTF.

This job is triggered to run daily.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: I175bdaee0d0360789c3c61051f2c56be8ba9ff89