tf-a-main: Add "aliases" for multijob subjobs
This is cosmetic measure which improves visibility into what actual build
happens / where error happened. Alias is shown in the job's subjob table,
in parens after the job name. E.g., while before it was:

tf-a-ci-gateway
tf-a-ci-gateway
[same job repeated a dozen of times w/o differentiation]

now it will be:

tf-a-ci-gateway (tf-l1-build-fvp tf-l2-build-fvp)
tf-a-ci-gateway (tf-l1-build-juno tf-l2-build-juno)

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I85220c35095e1c8f65dfe5bdfaecb444b46f8b5c